Updated Genetic
Genetic Updated to add support for inversion, permutation, 2-point crossover. 
  -These actions can be controlled by adjusting the associated probability rates. 
Population is now unique. 
Fixed bug in mutate-with-chance
If procedure for creating children results in a child already in population then a random child is created instead. Supporting functions for random child added in utility.
No need for population size to be in the thousands range for satisfying results.